# server/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Gym dashboard",
    "description": ""
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "tags": [],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/workout/": {
      "post": {
        "tags": [
          "CRUD"
        ],
        "description": "Create a new workout",
        "parameters": [
          {
            "name": "New workout",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/workout"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "get": {
        "tags": [
          "CRUD"
        ],
        "description": "Get all workouts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/workout/{id}": {
      "get": {
        "tags": [
          "CRUD"
        ],
        "description": "Get a workout",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "CRUD"
        ],
        "description": "Update a workout",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Updated workout",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/workout"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/workout-creation": {
      "post": {
        "tags": [
          "Processes"
        ],
        "description": "Initiate a workout creation.\r  Number of inputs must match template parameters",
        "parameters": [
          {
            "name": "Data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/workoutCreation"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "workout": {
      "type": "object",
      "properties": {
        "workoutName": {
          "type": "string",
          "example": "Test workout"
        },
        "status": {
          "type": "string",
          "example": "PENDING"
        },
        "excercises": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "excerciseName": {
                "type": "string",
                "example": "Test exercise"
              },
              "sets": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "reps": {
                      "type": "string",
                      "example": "10"
                    },
                    "load": {
                      "type": "string",
                      "example": "100 kg"
                    },
                    "time": {
                      "type": "number",
                      "example": 120
                    }
                  },
                  "required": [
                    "reps",
                    "load",
                    "time"
                  ]
                }
              }
            },
            "required": [
              "excerciseName",
              "sets"
            ]
          }
        }
      },
      "required": [
        "workoutName",
        "status",
        "excercises"
      ]
    },
    "workoutCreation": {
      "type": "object",
      "properties": {
        "template": {
          "type": "string",
          "example": "TEMPLATE_ID"
        },
        "inputs": {
          "type": "array",
          "example": [
            "VAR#1",
            "VAR#2"
          ],
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "template",
        "inputs"
      ]
    }
  }
}